Orders.xlsx file
Hello!
Changed the Excel filename. The content of the file (Ordes.xlsx) that exists right now in the repo is not equal to the suggestion in step 1. I can't upload a file to the repo, so I can't change that.
Anyhow, the repo file does not work for me. It throws an error on the date format.
Sorry for posting this after the first update, but I noticed this content error just now.
Best regards,
Rafael